feat(queries): 'useTemplateListQuery' 구현
