--- frontend/src/hooks/template/useTemplateListQuery.ts
@@ -1,10 +1,0 @@
-import { useQuery } from '@tanstack/react-query';
-
-import { QUERY_KEY, getTemplateList } from '@/api';
-import type { TemplateListResponse } from '@/types';
-
-export const useTemplateListQuery = () =>
-  useQuery<TemplateListResponse, Error>({
-    queryKey: [QUERY_KEY.TEMPLATE_LIST],
-    queryFn: getTemplateList,
-  });
