+++ frontend/src/queries/template/index.ts
@@ -1,0 +1,1 @@
+export { useTemplateListQuery } from './useTemplateListQuery';
